+++ samples/ProjectReference/Host/Program.cs
@@ -107,10 +107,7 @@ cloudShell.DefineResources(resources =>
 builder.Services
     .AddAspNetCoreProjectResourceType()
     .AddResourceModelGraphServices()
-    .AddReferenceProviderResourceManagerProjections()
-    .AddResourceModelGraphProcedureProvider(
-        ResourceModelResourceProvider.DefaultProviderId,
-        "Resource model");
+    .AddReferenceProviderResourceManagerIntegration();
 
 cloudShell
     .AddExtension<ResourceManagerExtension>()